+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - `MergeMaps`, a `Monoid` on `Map` formed by `Map#merge` 
 
+### Fixed
+- issue where certain ways to compose `Effect`s unintentionally nullified the effect 
+
 ## [3.2.0] - 2018-12-08
 ### Changed
 - ***Breaking Change***: `Difference` and `Intersection` no longer instances of `Semigroup` and moved to `functions.builtin.fn2` package